+++ src/services/about.rs
@@ -2,7 +2,7 @@
 
 use std::sync::Arc;
 
-use super::authorization::{self, ACTION};
+use super::authorization::{self, Action};
 use crate::errors::AuthError;
 use crate::models::user::UserId;
 
@@ -26,7 +26,7 @@ impl Service {
     /// * There is an error authorizing the action.
     pub async fn get_about_page(&self, maybe_user_id: Option<UserId>) -> Result<String, AuthError> {
         self.authorization_service
-            .authorize(ACTION::GetAboutPage, maybe_user_id)
+            .authorize(Action::GetAboutPage, maybe_user_id)
             .await?;
 
         let html = r#"
@@ -60,7 +60,7 @@ impl Service {
     /// * There is an error authorizing the action.
     pub async fn get_license_page(&self, maybe_user_id: Option<UserId>) -> Result<String, AuthError> {
         self.authorization_service
-            .authorize(ACTION::GetLicensePage, maybe_user_id)
+            .authorize(Action::GetLicensePage, maybe_user_id)
             .await?;
 
         let html = r#"