+++ src/tests/services/mod.rs
@@ -1,4 +1,5 @@
 mod authentication;
+mod authorization;
 mod hasher;
 mod settings;
 mod torrent_file;