+++ src/web/api/client/v1/contexts/user/responses.rs
@@ -20,6 +20,8 @@ pub struct LoggedInUserData {
     pub token: String,
     pub username: String,
     pub admin: bool,
+    #[serde(default)]
+    pub role: String,
 }
 
 #[derive(Deserialize)]
@@ -42,4 +44,6 @@ pub struct TokenRenewalData {
     pub token: String,
     pub username: String,
     pub admin: bool,
+    #[serde(default)]
+    pub role: String,
 }