+++ src/web/api/server/v1/contexts/user/responses.rs
@@ -25,6 +25,7 @@ pub struct TokenResponse {
     pub token: String,
     pub username: String,
     pub admin: bool,
+    pub role: String,
 }
 
 /// Response after successfully logging in a user.
@@ -32,8 +33,9 @@ pub fn logged_in_user(token: String, user_compact: UserCompact) -> Json<OkRespon
     Json(OkResponseData {
         data: TokenResponse {
             token,
-            username: user_compact.username,
-            admin: user_compact.administrator,
+            username: user_compact.username.clone(),
+            admin: user_compact.is_admin(),
+            role: user_compact.role,
         },
     })
 }
@@ -43,8 +45,9 @@ pub fn renewed_token(token: String, user_compact: UserCompact) -> Json<OkRespons
     Json(OkResponseData {
         data: TokenResponse {
             token,
-            username: user_compact.username,
-            admin: user_compact.administrator,
+            username: user_compact.username.clone(),
+            admin: user_compact.is_admin(),
+            role: user_compact.role,
         },
     })
 }